+++ pages/api/get-ingredients.js
@@ -1,41 +1,30 @@
 import { parse } from 'node-html-parser';
+import getBackend from './third-parties';
 
-const defaultScraper = {
-  getList(document) {
-    return []
-  },
-  regex: /(?<quantity>\d+)(?: )?(?:(?<unit>[a-zA-Z]{1,4})?) (?<ingredient>[\w| ]+)/
-}
+const unitMap = {
+  g: 'gram',
+  kg: 'kilogram',
+  tbsp: 'tablespoon',
+  ml: 'millilitre',
+  l: 'litre',
+  tsp: 'teaspoon'
+};
 
-const BBCGoodFood = {
-  getList(document) {
-    const headings = document.querySelectorAll('h1,h2,h3,h4');
-    const ingredientHeading = [...headings].filter(heading => (
-      heading.innerText.match(/ingredients?/ig)
-    ));
-    const list = ingredientHeading[0].parentNode.querySelector('ul, ol').querySelectorAll('li');
-    return Array.from(list).map(li => li.innerText);
-  },
-
-  regex: /(?<quantity>\d+)(?: )?(?:(?<unit>[a-zA-Z]{1,4})?) (?<ingredient>[\w| ]+)/
-}
-
-const hostnameMap = {
-  'www.bbcgoodfood.com': BBCGoodFood
-}
-
-function getList(html, hostname) {
+function getList(html, backend) {
   const document = parse(html);
-  return (hostnameMap[hostname] || defaultScraper).getList(document)
+  return backend.getList(document)
 }
 
-function parseList(list, hostname) {
+function parseList(list, backend) {
   return list.map(li => {
-    const result = li.match((hostnameMap[hostname] || defaultScraper).regex);
+    const result = li.match(backend.regex);
     const { quantity, unit, ingredient } = result?.groups || {};
     // TODO: handle missing ones? Tidy up the response
     return {
-      text: li, quantity, unit, ingredient
+      text: li,
+      ingredient,
+      quantity,
+      unit: unitMap[unit] || unit
     }
   })
 }
@@ -45,12 +34,13 @@ export default async function handler(req, res) {
   try {
     const { hostname } = new URL(url);
     const html = await(await fetch(url)).text();
-    const list = getList(html, hostname);
-    const ingredients = parseList(list, hostname);
+    const backend = getBackend(hostname);
+    const list = getList(html, backend);
+    const ingredients = parseList(list, backend);
     res.status(200).json(ingredients);
   } catch (e) {
     console.error(e);
-    res.status(500).json({ error: 'failed to load data' });
+    res.status(500).json({ error: e.message });
   }
 
 }
